refactor(core): split candidate attrs, facet place, ordinal, contracts
Extract identity series/mode and lineage/annotation helpers; assemble
render-model input type; chrome display types; trained scale types;
per-facet panel placement; single-panel margin reserve; finalize layer
vs domain contracts; ordinal train/result packing; bin frame packing;
bind-layer result assembly; lazy identity index. Facades keep import
paths stable. Add characterization tests for margin reserve and lazy index.
