+++ packages/core/src/pipeline/assemble-render-model-input.ts
@@ -1,0 +1,43 @@
+/**
+ * assembleRenderModel input contract.
+ */
+import type { TickFormatter } from "../layout/layout.js";
+import type { ScaleState } from "../scales/state.js";
+import type { PositionScale } from "../scales/train.js";
+import type { Scene } from "../scene.js";
+import type { CellValue, ColumnTable } from "../table.js";
+import type { CandidateStore } from "../candidate-store.js";
+import type { LineageStore } from "../identity.js";
+
+import type {
+  Advisory,
+  LayerBackend,
+  MappedField,
+  PipelineWarning,
+  ResolvedColorScale,
+  ScaleDomainSnapshot,
+} from "./types.js";
+
+export interface AssembleRenderModelInput {
+  scene: Scene;
+  xScale: PositionScale;
+  yScale: PositionScale;
+  color: ResolvedColorScale | null;
+  fill: ResolvedColorScale | null;
+  panelScales: { x: PositionScale; y: PositionScale }[];
+  colorState: ScaleState | null;
+  fillState: ScaleState | null;
+  warnings: PipelineWarning[];
+  advisories: Advisory[];
+  runId: number;
+  layerBackends: LayerBackend[];
+  layerFields: MappedField[][];
+  layerScaledConstants: ReadonlyArray<Readonly<Partial<Record<string, CellValue>>>>;
+  baselineDomains: ScaleDomainSnapshot;
+  effectiveDomains: ScaleDomainSnapshot;
+  lineage: LineageStore<number>;
+  candidates: CandidateStore;
+  formatX: TickFormatter | undefined;
+  formatY: TickFormatter | undefined;
+  table: ColumnTable;
+}
