+++ packages/core/src/pipeline/bind-layer-result.ts
@@ -1,0 +1,38 @@
+/**
+ * Assemble LayerBinding from resolved channel fields and extras.
+ */
+import type { LayerSpec } from "@ggsvelte/spec";
+
+import type { LayerBinding } from "./types.js";
+
+export function makeLayerBinding(input: {
+  layer: LayerSpec;
+  index: number;
+  xField: string | null;
+  yField: string | null;
+  yStatColumn: string | null;
+  yminField: string | null;
+  ymaxField: string | null;
+  color: LayerBinding["color"];
+  fill: LayerBinding["fill"];
+  labelField: string | null;
+  labelConstant: string | null;
+  weightField: string | null;
+  ruleForm: LayerBinding["ruleForm"];
+}): LayerBinding {
+  return {
+    layer: input.layer,
+    index: input.index,
+    xField: input.xField,
+    yField: input.yField,
+    yStatColumn: input.yStatColumn,
+    yminField: input.yminField,
+    ymaxField: input.ymaxField,
+    color: input.color,
+    fill: input.fill,
+    labelField: input.labelField,
+    labelConstant: input.labelConstant,
+    weightField: input.weightField,
+    ruleForm: input.ruleForm,
+  };
+}
