+++ packages/core/src/pipeline/frame-stats-bin.ts
@@ -6,10 +6,10 @@ import type { BarParams } from "@ggsvelte/spec";
 import { statBin } from "../stats/bin.js";
 import type { ColumnTable } from "../table.js";
 
-import { carriedColumns, emptyFrameExtras, removedStatWarning } from "./frame-helpers.js";
+import { carriedColumns, removedStatWarning } from "./frame-helpers.js";
+import { packBinLayerFrame } from "./frame-stats-bin-frame.js";
 import { makeColumnOf } from "./frame-stats-shared.js";
 import type { Advisory, LayerBinding, LayerFrame, PipelineWarning } from "./types.js";
-import { NO_ROW } from "./types.js";
 
 export function buildBinFrame(
   binding: LayerBinding,
@@ -43,27 +43,5 @@ export function buildBinFrame(
       howToOverride: `Set params.binwidth (preferred — a width meaningful for the data) or params.bins on layer ${index}.`,
     });
   }
-  const columns: Record<string, Float64Array> = {
-    count: result.count,
-    density: result.density,
-    ncount: result.ncount,
-    ndensity: result.ndensity,
-  };
-  const col = columnOf(result, null);
-  return {
-    binding,
-    table,
-    n: result.x.length,
-    xValues: null,
-    xNumeric: result.x,
-    yNumeric: columns[binding.yStatColumn ?? "count"] ?? result.count,
-    groups: result.groups,
-    rowIndex: Uint32Array.from({ length: result.x.length }, () => NO_ROW),
-    colorValues: col(binding.color.field),
-    fillValues: col(binding.fill.field),
-    labelValues: col(binding.labelField),
-    ...emptyFrameExtras(),
-    xmin: result.xmin,
-    xmax: result.xmax,
-  };
+  return packBinLayerFrame(binding, table, result, columnOf);
 }
