+++ packages/core/src/pipeline/panel-layout-single.ts
@@ -1,12 +1,13 @@
 /**
  * Single-panel layout placement with axis-title and legend reserves.
  */
-import type { LayoutTheme, Margins, TickFormatter } from "../layout/layout.js";
+import type { LayoutTheme, TickFormatter } from "../layout/layout.js";
 import { layout } from "../layout/layout.js";
 import type { TextMeasurer } from "../layout/measure.js";
 import type { PositionScale } from "../scales/train.js";
 
-import { LEGEND_EDGE_PAD, LEGEND_GAP, layoutDomain } from "./layout-helpers.js";
+import { layoutDomain } from "./layout-helpers.js";
+import { singlePanelMarginReserve } from "./panel-layout-single-reserve.js";
 import type { PanelPlacement } from "./panel-layout-types.js";
 
 export function placeSinglePanel(input: {
@@ -44,11 +45,6 @@ export function placeSinglePanel(input: {
     layoutTheme,
   } = input;
 
-  const reserve: Partial<Margins> = {
-    ...(hTitle !== "" && { bottom: axisTitleBand }),
-    ...(vTitle !== "" && { left: axisTitleBand }),
-    ...(legendWidth > 0 && { right: legendWidth + LEGEND_GAP + LEGEND_EDGE_PAD }),
-  };
   const layoutResult = layout({
     width: optionsWidth,
     height: layoutHeight,
@@ -57,7 +53,7 @@ export function placeSinglePanel(input: {
     ...(formatH !== undefined && { formatX: formatH }),
     ...(formatV !== undefined && { formatY: formatV }),
     measurer,
-    reserve,
+    reserve: singlePanelMarginReserve(hTitle, vTitle, axisTitleBand, legendWidth),
     theme: layoutTheme,
   });
   const margins = layoutResult.margins;
