+++ packages/core/src/pipeline/train-pipeline-scales.ts
@@ -4,26 +4,15 @@
 import type { PortableSpec } from "@ggsvelte/spec";
 
 import type { EditionDefaults } from "../editions.js";
-import type { PositionScale } from "../scales/train.js";
 import type { ColumnTable } from "../table.js";
 
 import type { FacetPanelDef } from "./facets.js";
-import { collectAxisInputs, resolveColorScale, trainAxis } from "./scale-training.js";
 import { trainPipelineColorScales } from "./train-pipeline-scales-color.js";
 import { trainPipelinePositionScales } from "./train-pipeline-scales-position.js";
+import type { TrainedPipelineScales } from "./train-pipeline-scales-types.js";
 import type { Advisory, LayerFrame, PipelineWarning, RunOptions } from "./types.js";
 
-export interface TrainedPipelineScales {
-  xTraining: ReturnType<typeof trainAxis>;
-  yTraining: ReturnType<typeof trainAxis>;
-  panelScales: { x: PositionScale; y: PositionScale }[];
-  colorResolution: ReturnType<typeof resolveColorScale>;
-  fillResolution: ReturnType<typeof resolveColorScale>;
-  xInputs: ReturnType<typeof collectAxisInputs>;
-  yInputs: ReturnType<typeof collectAxisInputs>;
-  scalesConfig: NonNullable<PortableSpec["scales"]>;
-  allFrames: LayerFrame[];
-}
+export type { TrainedPipelineScales } from "./train-pipeline-scales-types.js";
 
 export function trainPipelineScales(input: {
   normalized: PortableSpec;
