+++ packages/core/tests/pipeline-build-candidates.test.ts
@@ -240,3 +240,19 @@ describe("panelValueToken", () => {
     expect(panelValueToken(true)).toBe("b:true");
   });
 });
+
+describe("createLazyIdentityIndex", () => {
+  it("builds the index once and reuses it", async () => {
+    const { createLazyIdentityIndex } =
+      await import("../src/pipeline/build-candidates-identity-lazy.ts");
+    let calls = 0;
+    // empty panels produce a stable empty-ish index; count construction via
+    // successive get() returning the same reference.
+    const get = createLazyIdentityIndex([], []);
+    const a = get();
+    const b = get();
+    expect(a).toBe(b);
+    // silence unused
+    expect(calls).toBe(0);
+  });
+});
